Notebooks/chatbot_modules/requirements.txt
# Основные зависимости
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.32.0
streamlit-webrtc>=0.47.1  # Для записи аудио через WebRTC

# Для распознавания речи
SpeechRecognition>=3.10.0
PyAudio>=0.2.13  # Для работы с микрофоном
whisper>=1.0.0  # Опционально, для локального распознавания с помощью Whisper
pocketsphinx>=5.0.0  # Опционально, для локального распознавания без интернета

# Для поиска в интернете
google-search-results>=2.4.2  # Для Serper API